1. fix g++ warning 	dc_driv.cpp 	src_external/src_pdiag/Cblacs.h 	src_external/src_pdiag/GenELPA.cpp 	src_external/src_pdiag/blas_interface.h 	src_global/complexarray.cpp 	src_global/complexmatrix.h 	src_global/global_function.cpp 	src_global/intarray.cpp 	src_global/lapack_connector.h 	src_global/mathzone.cpp 	src_global/mathzone.h 	src_global/mathzone_add1.cpp 	src_global/matrix.h 	src_global/poission.cpp 	src_global/polint.cpp 	src_global/scalapack_connector.h 	src_global/sph_bessel_recursive-d1.cpp 	src_global/sph_bessel_recursive-d2.cpp 	src_global/timer.cpp 	src_global/timer.h 	src_lcao/Mulliken_Charge.cpp 	src_lcao/center2_orb-orb11.cpp 	src_lcao/diago_lcao_matrix.cpp 	src_lcao/gint_gamma_rho.cpp 	src_lcao/gint_gamma_vl.cpp 	src_lcao/gint_k_vl.cpp 	src_lcao/hs_matrix.cpp 	src_lcao/lcao_orbitals.cpp 	src_lcao/local_orbital_charge.cpp 	src_lcao/local_orbital_elec.cpp 	src_lcao/local_orbital_wfc.cpp 	src_lcao/make_gaunt_table.cpp 	src_lcao/make_vna_table.cpp 	src_lcao/numerical_orbital_lm.cpp 	src_lcao/numerical_vna_lm.cpp 	src_lcao/record_adj.cpp 	src_lcao/sltk_atom_input.cpp 	src_lcao/sltk_grid.cpp 	src_lcao/update_input.cpp 	src_lcao/use_hamilt_matrix.cpp 	src_lcao/use_overlap_matrix.cpp 	src_lcao/use_overlap_table.cpp 	src_lcao/ylm.cpp 	src_parallel/ft.cpp 	src_parallel/parallel_global.cpp 	src_parallel/parallel_kpoints.cpp 	src_pw/charge.cpp 	src_pw/charge_extra.cpp 	src_pw/charge_mixing.cpp 	src_pw/chi0_hilbert.cpp 	src_pw/chi0_standard.cpp 	src_pw/efield.cpp 	src_pw/energy.cpp 	src_pw/functional.cpp 	src_pw/gga_pw.cpp 	src_pw/hamilt_pw.cpp 	src_pw/myfunc2.cpp 	src_pw/myfunc5.cpp 	src_pw/numerical_basis.cpp 	src_pw/potential.cpp 	src_pw/pseudo_us.cpp 	src_pw/pseudopot_cell_vnl.cpp 	src_pw/pseudopot_upf.cpp 	src_pw/pw_basis.cpp 	src_pw/pw_complement.cpp 	src_pw/threshold_elec.cpp 	src_pw/toWannier90.cpp 	src_pw/unitcell_pseudo.cpp 	src_pw/vdwd2.cpp 	src_pw/vdwd3.cpp 	src_pw/wavefunc_in_pw.cpp 	src_pw/wavefunc_in_pw.h 	src_pw/xc_functional.cpp

+++ ABACUS.1.0.0/source/src_global/sph_bessel_recursive-d2.cpp
@@ -40,7 +40,7 @@ const vector<vector<vector<double>>> & Sph_Bessel_Recursive::D2::cal_jlx( const 
 
 void Sph_Bessel_Recursive::D2::cal_jlx_0( const int l_size, const size_t ix1_size, const size_t ix2_size )
 {
-	if(jlx.size()<l_size)
+	if(jlx.size() < static_cast<size_t>(l_size))
 		jlx.resize(l_size);
 	
 	for( int l=0; l!=l_size; ++l )
@@ -65,7 +65,7 @@ void Sph_Bessel_Recursive::D2::cal_jlx_0( const int l_size, const size_t ix1_siz
 
 void Sph_Bessel_Recursive::D2::cal_jlx_smallx( const int l_size, const size_t ix1_size, const size_t ix2_size )
 {
-	if(jlx.size()<l_size)
+	if(jlx.size() < static_cast<size_t>(l_size))
 		jlx.resize(l_size);
 	
 	for( int l=0; l!=l_size; ++l )
@@ -92,7 +92,7 @@ void Sph_Bessel_Recursive::D2::cal_jlx_smallx( const int l_size, const size_t ix
 
 void Sph_Bessel_Recursive::D2::cal_jlx_recursive( const int l_size, const size_t ix1_size, const size_t ix2_size )
 {
-	if(jlx.size()<l_size)
+	if(jlx.size() < static_cast<size_t>(l_size))
 		jlx.resize(l_size);
 	
 	for( int l=0; l!=l_size; ++l )
